Inscription MIT/Code/Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7
Pillow==10.0.1
qrcode==7.4.2
python-dotenv==1.0.0
psycopg2==2.9.10
